Restructured project to hide rlbot interface from user.
