+++ rlbot-src/inc/botmanager.h
