+++ rlbot-src/inc/errorcodes.h
