+++ rlbot-src/inc/namedrenderer.h
