+++ rlbot-src/inc/statesetting.h
