+++ rlbot-src/lib/inc/flatbuffers/registry.h
