+++ rlbot-src/lib/inc/flatbuffers/stl_emulation.h
