+++ rlbot-src/lib/inc/flatbuffers/util.h
@@ -32,8 +32,8 @@
 #define NOMINMAX
 #endif
 #include <direct.h>
-#include <winbase.h>
-#include <windows.h>  // Must be included before <direct.h>
+#include <WinBase.h>
+#include <Windows.h>  // Must be included before <direct.h>
 #undef interface      // This is also important because of reasons
 #else
 #include <limits.h>
