+++ rlbot-src/src/botmanager.cc
