+++ rlbot-src/src/namedrenderer.cc
