+++ rlbot-src/src/renderer.cc
